This commit shows that latency is good enough
I was having weird latency, but I think it's because I was introducing
some linearity in the sending of the note value and trigger where
one was waiting for the other. When the midi messages are sent concurrently
the latency is imperceptible.

There may have been added latency with my midi message broadcasting setup too?